Trait fix_hidden_lifetime_bug::Captures

pub trait Captures<'__> { }

The main hack allowing to mention extra lifetime parameters in an impl Trait type without expressing an outlives relationship.

As mentioned in the issues / GitHub comments referenced in the crate docs, this is the main tool to solve the hidden lifetime bugs.

Indeed, it is even officially present among the internal helper types that the very Rust compiler uses for itself: ::rustc_data_structures

If the above link were not to show the rustc_data_structures definition, then feel free to follow the following direct commit-tagged link